use real db migrations for sql module tests
